refactor: addons toggle to only show when transactionList exists
