+++ src/components/CustomerQuota/ItemsSelection/ItemContent.tsx
@@ -1,12 +1,11 @@
 import React, { FunctionComponent } from "react";
-import { View, StyleSheet, GestureResponderEvent } from "react-native";
+import { View, StyleSheet } from "react-native";
 import { AppText } from "../../Layout/AppText";
 import { CampaignPolicy } from "../../../types";
 import { ItemMaxUnitLabel } from "./ItemMaxUnitLabel";
 import { fontSize, color } from "../../../common/styles";
 import { sharedStyles } from "./sharedStyles";
 import { useTranslate } from "../../../hooks/useTranslate/useTranslate";
-import { ShowAddonsToggle, DescriptionAlertTypes } from "./ShowAddonsToggle";
 
 const styles = StyleSheet.create({
   name: {
@@ -25,21 +24,15 @@ const styles = StyleSheet.create({
 export const ItemContent: FunctionComponent<{
   name: CampaignPolicy["name"];
   description: CampaignPolicy["description"];
-  descriptionAlert?: DescriptionAlertTypes;
   unit: CampaignPolicy["quantity"]["unit"];
   maxQuantity: number;
   accessibilityLabel?: string;
-  showAddonsToggle?: (e: GestureResponderEvent) => void;
-  showAddons?: boolean;
 }> = ({
   name,
   description,
-  descriptionAlert,
   unit,
   maxQuantity,
   accessibilityLabel = "item-content",
-  showAddonsToggle,
-  showAddons,
 }) => {
   const { c13nt } = useTranslate();
   const tDescription = c13nt(description ?? "");
@@ -61,13 +54,6 @@ export const ItemContent: FunctionComponent<{
         <AppText style={sharedStyles.maxQuantityLabel}>
           <ItemMaxUnitLabel unit={unit} maxQuantity={maxQuantity} />
         </AppText>
-      )}
-      {descriptionAlert && descriptionAlert.length > 0 && (
-        <ShowAddonsToggle
-          descriptionAlert={descriptionAlert}
-          toggleIsShowAddons={showAddonsToggle}
-          isShowAddons={showAddons}
-        />
       )}
     </View>
   );
